moving theme hook code to a hook file
