+++ web/components/Label.js
@@ -1,4 +1,4 @@
-import { styled } from "../lib/theme.js";
+import { styled } from "./hook-theme.js";
 
 const Label = styled('span', {
   fontSize: '0.75rem',
